transfer-old-data/requirements.txt
pylint
pytest
pandas
boto3
python-dotenv
sqlalchemy
pymssql
fsspec
s3fs